% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{build_detail_overlay}
\alias{build_detail_overlay}
\title{Build Detail Overlay for Vegbank UI}
\usage{
build_detail_overlay()
}
\value{
A Shiny tag representing the detail overlay.
}
\description{
Constructs the overlay panel that displays detailed plot information.
}
\keyword{internal}
