% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{build_details_view}
\alias{build_details_view}
\title{Build Details View}
\usage{
build_details_view(selected_data)
}
\arguments{
\item{selected_data}{A data row representing the selected plot.}
}
\value{
A list of Shiny UI outputs.
}
\description{
Constructs a list of Shiny UI outputs for displaying detailed plot information.
}
\keyword{internal}
