% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{build_navbar}
\alias{build_navbar}
\title{Build Navigation Bar for Vegbank UI}
\usage{
build_navbar()
}
\value{
A Shiny tag list representing the navigation bar.
}
\description{
Constructs and returns the navigation bar to be used in the UI.
}
\keyword{internal}
