% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_visualizations.R
\name{build_plot_heatmap}
\alias{build_plot_heatmap}
\title{Build Plot Heatmap}
\usage{
build_plot_heatmap(data)
}
\arguments{
\item{data}{Data frame containing longitude and latitude.}
}
\value{
A ggplot object.
}
\description{
Constructs a heatmap (using ggplot2) of plot locations using density estimation.
}
\keyword{internal}
