% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{build_taxa_list}
\alias{build_taxa_list}
\title{Build Taxa List}
\usage{
build_taxa_list(data_row)
}
\arguments{
\item{data_row}{A row of data that contains taxa information.}
}
\value{
A character string containing HTML markup.
}
\description{
Generates an HTML ordered list of taxa from a data row.
}
\keyword{internal}
