% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detail_view.R
\name{build_taxon_details_view}
\alias{build_taxon_details_view}
\title{Build Taxon Details View}
\usage{
build_taxon_details_view(taxon_data)
}
\arguments{
\item{taxon_data}{A data row representing the selected taxon.}
}
\value{
A list of Shiny UI outputs.
}
\description{
Constructs a list of Shiny UI outputs for displaying detailed taxon information.
}
\keyword{internal}
