% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{build_top10_barchart}
\alias{build_top10_barchart}
\title{Build Top Ten Bar Chart}
\usage{
build_top10_barchart(data, column, xlab, color)
}
\arguments{
\item{data}{Data frame containing the data.}

\item{column}{Name of the column to summarize.}

\item{xlab}{Label for the x-axis.}

\item{color}{Bar fill color.}
}
\value{
A ggplot object.
}
\description{
Constructs a ggplot bar chart based on the top 10 counts for a specified column.
}
\keyword{internal}
