% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\docType{data}
\name{custom_theme}
\alias{custom_theme}
\title{Custom Bootstrap Theme for Vegbank Web Application}
\format{
An object of class \code{bs_theme_with_preset} (inherits from \code{bs_version_5}, \code{bs_theme}, \code{sass_bundle}) of length 1.
}
\usage{
custom_theme
}
\value{
A Bootstrap theme object.
}
\description{
Defines a bslib theme with custom rules.
}
\keyword{internal}
