% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_handler.R
\name{ph_inc_progress}
\alias{ph_inc_progress}
\title{Increment progress in a test-safe way}
\usage{
ph_inc_progress(amount, detail = NULL)
}
\arguments{
\item{amount}{The amount to increment}

\item{detail}{The detail message}
}
\value{
Invisible NULL
}
\description{
Increment progress in a test-safe way
}
\keyword{internal}
