% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_handler.R
\name{ph_is_in_shiny_session}
\alias{ph_is_in_shiny_session}
\title{Check if current environment is an interactive Shiny session}
\usage{
ph_is_in_shiny_session()
}
\value{
Boolean indicating if the current environment is an interactive Shiny session
}
\description{
Check if current environment is an interactive Shiny session
}
\keyword{internal}
