% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_handler.R
\name{ph_is_in_test_env}
\alias{ph_is_in_test_env}
\title{Progress and Notification Handler Module}
\usage{
ph_is_in_test_env()
}
\value{
Boolean indicating if the current environment is a test
}
\description{
Provides functions for handling progress indicators and notifications in a test-safe way.
}
\keyword{Check}
\keyword{a}
\keyword{current}
\keyword{environment}
\keyword{if}
\keyword{internal}
\keyword{is}
\keyword{test}
