% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\docType{data}
\name{plot_map}
\alias{plot_map}
\title{Map Functions Module}
\format{
An object of class \code{list} of length 3.
}
\usage{
plot_map
}
\description{
Provides functions for creating and manipulating leaflet maps.
}
\keyword{internal}
