% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_table.R
\docType{data}
\name{plot_table}
\alias{plot_table}
\title{Table Data Functions Module}
\format{
An object of class \code{list} of length 2.
}
\usage{
plot_table
}
\description{
Provides functions for creating and manipulating the plot data table.
}
\keyword{internal}
