% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run Vegbank Web Application}
\usage{
run_app(options = list())
}
\arguments{
\item{options}{A list of options to pass to the Shiny application.
Default is an empty list.}
}
\value{
A \strong{Shiny application} object.
}
\description{
Launches the Vegbank Shiny application.
}
\details{
Running the application requires:
\itemize{
\item R (version 4.0 or higher) installed on the local machine
\item The \code{vegbankweb} package and its dependencies installed in the R environment
\item R_CONFIG_FILE environment variable set to the absolute path of a valid configuration file
\item An active internet connection to access the Vegbank API
\item A modern web browser to view the user interface
}
}
\concept{\strong{Standalone Application}}
