% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Shiny Server for Vegbank Web Application}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}
}
\value{
Called for its side effects.
}
\description{
Initializes server-side functionality for data visualization and interactivity.
}
