% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detail_view.R
\name{show_detail_view}
\alias{show_detail_view}
\title{Detail View Functions}
\usage{
show_detail_view(detail_type, accession_code, output, session)
}
\arguments{
\item{detail_type}{Type of detail to show ("plot", "community", or "taxon-observation")}

\item{accession_code}{The accession code to fetch details for}

\item{output}{The Shiny output object}

\item{session}{The Shiny session object}
}
\value{
Boolean indicating success or failure
}
\description{
Functions for building the detail view panel.
}
\keyword{Detail}
\keyword{Generic}
\keyword{Show}
\keyword{View}
\keyword{and}
\keyword{details}
\keyword{display}
\keyword{fetch}
\keyword{function}
\keyword{in}
\keyword{internal}
\keyword{overlay.}
\keyword{the}
\keyword{to}
