% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui}
\alias{ui}
\title{Shiny UI for Vegbank Web Application}
\usage{
ui(req)
}
\arguments{
\item{req}{A Shiny request object.}
}
\value{
A Shiny tag list.
}
\description{
Constructs the user interface for browsing vegetation plot data.
}
