% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\docType{data}
\name{veg_bank_api}
\alias{veg_bank_api}
\title{VegBank API Client Module}
\format{
An object of class \code{list} of length 4.
}
\usage{
veg_bank_api
}
\description{
Provides functions for interacting with the VegBank API.
}
\keyword{internal}
