% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detail_view.R
\name{build_details_view}
\alias{build_details_view}
\title{Detail View Functions}
\usage{
build_details_view(selected_data)
}
\arguments{
\item{selected_data}{A data row representing the selected plot.}
}
\value{
A list of Shiny UI outputs.
}
\description{
Functions for building the detail view panel.
}
\keyword{Build}
\keyword{Constructs}
\keyword{Details}
\keyword{Shiny}
\keyword{UI}
\keyword{View}
\keyword{a}
\keyword{detailed}
\keyword{displaying}
\keyword{for}
\keyword{information.}
\keyword{internal}
\keyword{list}
\keyword{of}
\keyword{outputs}
\keyword{plot}
