% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_visualizations.R
\name{build_most_recent_date_list}
\alias{build_most_recent_date_list}
\title{Build Most Recent Date List}
\usage{
build_most_recent_date_list(data, n = 16, date_column = "obsdateentered")
}
\arguments{
\item{data}{Data frame containing date information.}

\item{n}{Maximum number of dates to display.}

\item{date_column}{Name of the date column}
}
\value{
A Shiny UI tag containing an unordered list.
}
\description{
Creates a list of the most recent dates from the data.
}
\keyword{internal}
