% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_visualizations.R
\name{build_pie_chart}
\alias{build_pie_chart}
\title{Build Pie Chart}
\usage{
build_pie_chart(
  data,
  column,
  palette = c("#a1d99b", "#31a354"),
  label_top_n = 4
)
}
\arguments{
\item{data}{Data frame containing the data.}

\item{column}{The column name used for pie chart slices.}

\item{palette}{Color palette.}

\item{label_top_n}{Number of top labels to display.}
}
\value{
A plotly pie chart.
}
\description{
Constructs an interactive pie chart using plotly.
}
\keyword{internal}
