% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{build_taxa_list}
\alias{build_taxa_list}
\title{UI Helper Functions}
\usage{
build_taxa_list(data_row)
}
\arguments{
\item{data_row}{A row of data that contains taxa information.}
}
\value{
A character string containing HTML markup.
}
\description{
Functions for generating UI components and HTML elements.
}
\keyword{Build}
\keyword{Generates}
\keyword{HTML}
\keyword{List}
\keyword{Taxa}
\keyword{a}
\keyword{an}
\keyword{data}
\keyword{from}
\keyword{internal}
\keyword{list}
\keyword{of}
\keyword{ordered}
\keyword{row.}
\keyword{taxa}
