% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_visualiztions.R
\name{build_top10_barchart}
\alias{build_top10_barchart}
\title{Visualization Functions}
\usage{
build_top10_barchart(data, column, xlab, color)
}
\arguments{
\item{data}{Data frame containing the data.}

\item{column}{Name of the column to summarize.}

\item{xlab}{Label for the x-axis.}

\item{color}{Bar fill color.}
}
\value{
A ggplot object.
}
\description{
Functions for creating various visualizations of VegBank data.
}
\keyword{10}
\keyword{Bar}
\keyword{Build}
\keyword{Chart}
\keyword{Constructs}
\keyword{Ten}
\keyword{Top}
\keyword{a}
\keyword{bar}
\keyword{based}
\keyword{chart}
\keyword{column.}
\keyword{counts}
\keyword{for}
\keyword{ggplot}
\keyword{internal}
\keyword{on}
\keyword{specified}
\keyword{the}
\keyword{top}
