% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run Vegbank Web Application}
\usage{
run_app(options = list())
}
\value{
A \strong{Shiny application} object.
}
\description{
Launches the Vegbank Shiny application.
}
\section{Details}{

The application uses:
\itemize{
\item \strong{UI}: Defined in \code{\link{ui.R}}.
\item \strong{Server Logic}: Defined in \code{\link{server.R}}.
}
}

\concept{\strong{Standalone Application}}
