% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run Vegbank Web Application}
\usage{
run_app(options = list())
}
\arguments{
\item{options}{A list of options to pass to the Shiny application.
Default is an empty list.}
}
\value{
A \strong{Shiny application} object.
}
\description{
Launches the Vegbank Shiny application.
}
\concept{\strong{Standalone Application}}
